requirements.txt
tensorflow
keras-tuner
scikit-learn
matplotlib
seaborn
shutils
PyYAML
dvc